Consolidated framework thread_local into one package
Moved a thread_local from FWCore/Framework to FWCore/ServiceRegistry
since it was the true dependencies and it makes FWCore/Framework have
no thread_local and FWCore/ServiceRegistry already has a thread_local.
This reduces the number of libraries with thread_locals which is a
limited resource for linux.
