+++ FWCore/ServiceRegistry/src/CurrentModuleOnThread.cc
@@ -1,3 +1,3 @@
-#include "FWCore/Framework/interface/CurrentModuleOnThread.h"
+#include "FWCore/ServiceRegistry/interface/CurrentModuleOnThread.h"
 
 thread_local edm::ModuleCallingContext const* edm::CurrentModuleOnThread::currentModuleOnThread_ = nullptr;
